requirements.txt
python-dotenv
pyyaml
paho-mqtt
adafruit-blinka
adafruit-circuitpython-lis3dh
adafruit-circuitpython-dht
adafruit-circuitpython-bme280
adafruit-circuitpython-mcp3xxx
adafruit-circuitpython-logging
